+++ client/src/App.js
@@ -19,6 +19,7 @@ import './App.css';
 import PostLogIn from './Components/NavBar/Authentication/PostLogIn';
 import CreateExperiences from './Components/CreateExperiences/CreateExperiences';
 import Bookings from './Components/Profile/AdminProfile/Bookings';
+import Users from './Components/Profile/AdminProfile/Users';
 
 function App() {
   return (
@@ -40,6 +41,7 @@ function App() {
         <Route exact path= '/editProfile' component={ ProfileToEdit } />
         <Route exact path= '/AdminProfile' component={ AdminProfile } />
         <Route exact path= '/AdminProfile/bookings' component={ Bookings } />
+        <Route exact path= '/AdminProfile/users' component={ Users } />
         <Route exact path= '/experience/create' component={ CreateExperiences } />
         <Route path='*' component={ Error } />
       </Switch>
