+++ client/src/Components/Profile/AdminProfile/AdminProfile.jsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from "react";
+import React, { useState } from "react";
 import { useHistory } from "react-router-dom";
 import { useDispatch, useSelector } from "react-redux";
 import { useAuth0 } from "@auth0/auth0-react";
@@ -23,8 +23,7 @@ import { getUser, updateUser } from "../../../Redux/actions";
 import background from "../../../image/fondo_huellas.png";
 import { AdminSearchBar } from "./SearchBar/AdminSearchBar";
 // import RenderOneImage from "../../Cloudinary/renderOneImage";
-import axios from "axios";
-import { DataGrid } from "@mui/x-data-grid";
+
 
 export const AdminProfile = () => {
   const dispatch = useDispatch();
@@ -39,12 +38,7 @@ export const AdminProfile = () => {
   });
   const [errors, setErrors] = useState({});
   const [photo, setPhoto] = useState(undefined);
-  const [users, setUsers] = useState("");
   const userToModify = useSelector((state) => state.user);
-
-  useEffect(() => {
-    axios("/user/getAll").then((res) => setUsers(res.data));
-  }, []);
 
   if (isLoading) return <Loading />;
   if (!user) history.goBack();
@@ -90,27 +84,13 @@ export const AdminProfile = () => {
       !(input.firstName || input.lastName || input.cellphone || input.userName)
     : true;
 
-  const rowsUsers = users.map((user) => {
-    return {
-      id: user.idUser,
-      col1: user.firstName,
-      col2: user.lastName,
-      col3: user.email,
-      col4: user.isAdmin ? "Administrador" : "Usuario",
-    };
-  });
-  const columnsUsers = [
-    { field: "id", headerName: "ID", width: 50 },
-    { field: "col1", headerName: "Nombre", width: 100 },
-    { field: "col2", headerName: "Apellido", width: 100 },
-    { field: "col3", headerName: "Email", width: 100 },
-    { field: "col4", headerName: "Status", width: 100 },
-  ];
-
   const seeBookings = () => {
-    history.push('/adminprofile/bookings')
-  }
+    history.push("/adminprofile/bookings");
+  };
 
+  const seeUsers = () => {
+    history.push("/adminprofile/users");
+  };
 
   return (
     <section className={s.allPage}>
@@ -123,14 +103,13 @@ export const AdminProfile = () => {
       <div className={s.bookings}>
         <h1>Reservas</h1>
         <Button
-              variant="contained"
-              color="success"
-              className={s.btnHome}
-              onClick={seeBookings}
-            >
-              Ver reservas
-            </Button>
-        
+          variant="contained"
+          color="success"
+          className={s.btnBook}
+          onClick={seeBookings}
+        >
+          Ver reservas
+        </Button>
       </div>
       <div className={s.container}>
         <h3 className={s.titulo}>Modificar datos de usuarios</h3>
@@ -238,14 +217,14 @@ export const AdminProfile = () => {
       </div>
       <div className={s.users}>
         <h1>Usuarios</h1>
-        <DataGrid
-          rows={rowsUsers}
-          columns={columnsUsers}
-          pageSize={8}
-          rowsPerPageOptions={[5]}
-          checkboxSelection
-          get
-        />
+        <Button
+          variant="contained"
+          color="success"
+          className={s.btnBook}
+          onClick={seeUsers}
+        >
+          Ver usuarios
+        </Button>
       </div>
 
       <img src={background} alt="fondo" className={s.background} />
