+++ client/src/Components/Profile/AdminProfile/Users.module.css
@@ -1,12 +1,12 @@
-.bookings{
+.users{
     width: 100%;
-    height: calc(100vh - 11rem);
+    height: calc(100vh - 6rem);
 }
 
 .goBack{
     position: absolute;
     top: 11rem;
-    left: 17rem;
+    left: 23rem;
 }
 .goBack svg{
     width: 3rem;
@@ -15,12 +15,12 @@
 }
 
 .list{
-    width: 65rem;
-    height: 23rem!important;
+    width: 53rem;
+    height: 32.9rem!important;
     margin-left: auto;
     margin-right: auto;
     background-color: #c1c1c1;
-    margin-top: 5rem;
+    margin-top: 5rem
 }
 
 .list div div div{
